Create User Manual for Application Workflow for Users
